+++ task/__init__.py
@@ -1,0 +1,1 @@
+from .bot_control import start, stop, status  # noqa: F401
